:art: cleaned up a lot by splitting index
